testing jacoco#2

+++ src/main/java/JUnit/tests/components/CPULimitTest.java
@@ -10,6 +10,6 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 public @interface CPULimitTest {
 	
-	long limit() default 75;
+	double limit() default 75;
 	
 }
